Add Controllers and PlayerShipController
Add controller components, which control how an entity acts. Add ship
controller, which adds several methods to control a ship. Add player
ship controller, which uses keyboard controls to move the ship.

+++ SeaGame/src/Game.h
@@ -7,7 +7,7 @@ class Game
 {
 public:
 	Game();
-	void update(unsigned int delta);
+	void update(double delta);
 	void render(sf::RenderWindow& w);
 
 private:
